requirements.txt
sdkit==1.0.47
ipywidgets==8.0.4
pynvml==11.5.0
xformers==0.0.16
python-dotenv==1.0.0
Flask==2.0.2
jsonschema==4.17.3
Pillow==9.4.0
gunicorn==20.1.0
urllib3>=1.26.5
chardet>=4.0.0
